[Feature] 의견 등록 시 이미지 업로드 기능 추가

+++ src/main/java/com/jisungin/application/talkroom/TalkRoomService.java
@@ -138,7 +138,8 @@ public class TalkRoomService {
         }
 
         if (request.getRemoveImage() != null && !request.getRemoveImage().isEmpty()) {
-            request.getRemoveImage().stream().map(s -> talkRoomImageRepository.findByTalkRoomAndImageUrl(talkRoom, s))
+            request.getRemoveImage().stream()
+                    .map(url -> talkRoomImageRepository.findByTalkRoomAndImageUrl(talkRoom, url))
                     .forEach(talkRoomImageRepository::deleteAll);
         }
     }
@@ -158,7 +159,7 @@ public class TalkRoomService {
         commentRepository.findByTalkRoom(talkRoom).ifPresent(commentRepository::delete);
 
         List<TalkRoomImage> images = talkRoomImageRepository.findByTalkRoom(talkRoom);
-        if (images != null) {
+        if (images != null && !images.isEmpty()) {
             talkRoomImageRepository.deleteAll(images);
         }
         talkRoomRoleRepository.deleteAllByTalkRoom(talkRoom);
